Trying to fix bugs around empty results in privateNotes cache (incomplete)

+++ static/js/s2.js
@@ -4870,6 +4870,9 @@ var AddNotePanel = React.createClass({
     this.props.setConnectionsMode("Tools");
   },
   deleteNote: function deleteNote() {
+    if (!confirm("Are you sure you want to delete this note?")) {
+      return;
+    }
     var url = "/api/notes/" + this.props.noteId;
     $.ajax({
       type: "delete",
